+++ packages/sheets-ui/src/services/selection/base-selection-render.service.ts
@@ -175,7 +175,7 @@ export class BaseSelectionRenderService extends Disposable implements ISheetSele
 
     readonly interceptor = new InterceptorManager({ RANGE_MOVE_PERMISSION_CHECK, RANGE_FILL_PERMISSION_CHECK });
 
-    private _escapeShortcutDisposable: Nullable<IDisposable> = null;
+    protected _escapeShortcutDisposable: Nullable<IDisposable> = null;
 
     constructor(
         protected readonly _injector: Injector,
@@ -255,8 +255,7 @@ export class BaseSelectionRenderService extends Disposable implements ISheetSele
     }
 
     newSelectionControl(scene: Scene, _rangeType: RANGE_TYPE, skeleton: SpreadsheetSkeleton): SelectionControl {
-        const selectionControls = this.getSelectionControls();
-        const zIndex = selectionControls.length;
+        const zIndex = this.getSelectionControls().length;
         const { rowHeaderWidth, columnHeaderHeight } = skeleton;
         const control = new SelectionControl(scene, zIndex, this._themeService, this._isHeaderHighlight, {
             rowHeaderWidth,
@@ -337,11 +336,14 @@ export class BaseSelectionRenderService extends Disposable implements ISheetSele
         });
     }
 
+    /**
+     * @TODO lumixraku DO NOT expose private props.
+     */
     getSelectionControls(): SelectionControl[] {
         return this._selectionControls;
     }
 
-    protected _clearSelectionControls(): void {
+    protected _clearAllSelectionControls(): void {
         const allSelectionControls = this._selectionControls;
         for (const control of allSelectionControls) {
             control.dispose();
@@ -398,7 +400,7 @@ export class BaseSelectionRenderService extends Disposable implements ISheetSele
     }
 
     /**
-     * get active selection control
+     * get active(actually last) selection control
      * @returns T extends SelectionControl
      */
     getActiveSelectionControl<T extends SelectionControl = SelectionControl>(): Nullable<T> {
@@ -418,152 +420,20 @@ export class BaseSelectionRenderService extends Disposable implements ISheetSele
     }
 
     /**
-     * Clear existed selections.
+     * Clear existed selections by workbookSelections.selectionMoveEnd$
      */
     protected _reset(): void {
-        this._clearSelectionControls();
+        this._clearAllSelectionControls();
         this._downObserver?.unsubscribe();
         this._downObserver = null;
     }
 
-    resetAndEndSelection(): void {
-        this.endSelection();
-        this._reset();
-    }
+    // resetAndEndSelection(): void {
+    //     this.endSelection();
+    //     this._reset();
+    // }
 
     // TODO: @wzhudev: refactor the method to make it more readable
-
-    /**
-     * Handle pointer down event, then trigger selectionMoveStart$.
-     * @param evt
-     * @param _zIndex
-     * @param rangeType
-     * @param viewport
-     * @param scrollTimerType
-     */
-    // eslint-disable-next-line max-lines-per-function, complexity
-    protected _onPointerDown(
-        evt: IPointerEvent | IMouseEvent,
-        _zIndex = 0,
-        rangeType: RANGE_TYPE = RANGE_TYPE.NORMAL,
-        viewport: Nullable<Viewport>,
-        scrollTimerType: ScrollTimerType = ScrollTimerType.ALL
-    ): void {
-        this._rangeType = rangeType;
-        // this._shouldDetectMergedCells = rangeType === RANGE_TYPE.NORMAL;
-
-        const skeleton = this._skeleton;
-        const scene = this._scene;
-        if (!scene || !skeleton) {
-            return;
-        }
-
-        if (viewport) {
-            this._activeViewport = viewport;
-        }
-
-        const { offsetX: evtOffsetX, offsetY: evtOffsetY } = evt;
-        const viewportMain = scene.getViewport(SHEET_VIEWPORT_KEY.VIEW_MAIN);
-        if (!viewportMain) return;
-        const relativeCoords = scene.getRelativeToViewportCoord(Vector2.FromArray([evtOffsetX, evtOffsetY]));
-
-        let { x: viewportPosX, y: viewportPosY } = relativeCoords;
-        this._startViewportPosX = viewportPosX;
-        this._startViewportPosY = viewportPosY;
-
-        const scrollXY = scene.getVpScrollXYInfoByPosToVp(relativeCoords);
-        const { scaleX, scaleY } = scene.getAncestorScale();
-        const cursorCellRangeInfo = this._getSelectRangeWithCoordByOffset(viewportPosX, viewportPosY, scaleX, scaleY, scrollXY);
-        if (!cursorCellRangeInfo) return;
-
-        const { rangeWithCoord: cursorCellRange, primaryWithCoord: primaryCursorCellRange } = cursorCellRangeInfo;
-        const cursorCellRangeWithRangeType: IRangeWithCoord = { ...cursorCellRange, rangeType };
-        this._startRangeWhenPointerDown = { ...cursorCellRange, rangeType };
-
-        let activeSelectionControl: Nullable<SelectionControl> = this.getActiveSelectionControl();
-        const curControls = this.getSelectionControls();
-        for (const control of curControls) {
-            // right click
-            if (evt.button === 2 && control.model.isInclude(cursorCellRangeWithRangeType)) {
-                activeSelectionControl = control;
-                return;
-            }
-            // Click to an existing selection
-            if (control.model.isEqual(cursorCellRangeWithRangeType)) {
-                activeSelectionControl = control;
-                break;
-            }
-
-            // There can only be one highlighted cell, so clear the highlighted cell of the existing selection
-            if (!evt.shiftKey) {
-                control.clearHighlight();
-            }
-        }
-
-        this._checkClearPreviousControls(evt);
-
-        const currentCell = activeSelectionControl?.model.currentCell;
-        const expandByShiftKey = evt.shiftKey && currentCell;
-        const remainLastEnable = this._remainLastEnabled &&
-            !evt.ctrlKey &&
-            !evt.shiftKey &&
-            !this._skipLastEnabled &&
-            !this._singleSelectionEnabled;
-
-        //#region update selection control
-        if (expandByShiftKey && currentCell) {
-            // Perform pointer down selection.
-            this._performSelectionByTwoCells(
-                currentCell,
-                cursorCellRangeWithRangeType,
-                skeleton,
-                rangeType,
-                activeSelectionControl! // Get updated in this method
-            );
-        } else if (remainLastEnable && activeSelectionControl) {
-            // Supports the formula ref text selection feature,
-            // under the condition of preserving all previous selections, it modifies the position of the latest selection.
-
-            activeSelectionControl.updateRange(cursorCellRangeWithRangeType, primaryCursorCellRange);
-        } else {
-            // Create new control as default
-            activeSelectionControl = this.newSelectionControl(scene, rangeType, skeleton);
-
-            activeSelectionControl.updateRange(cursorCellRangeWithRangeType, primaryCursorCellRange);
-        }
-        //#endregion
-
-        this._selectionMoveStart$.next(this.getSelectionDataWithStyle());
-
-        scene.disableObjectsEvent();
-        this._clearUpdatingListeners();
-        this._addEndingListeners();
-
-        scene.getTransformer()?.clearSelectedObjects();
-
-        scene.getEngine()?.setRemainCapture();
-
-        if (rangeType === RANGE_TYPE.ROW || rangeType === RANGE_TYPE.COLUMN) {
-            if (rangeType === RANGE_TYPE.ROW) {
-                viewportPosX = 0;
-            } else if (rangeType === RANGE_TYPE.COLUMN) {
-                viewportPosY = 0;
-            }
-            // TODO @lumixraku. This is so bad! There should be a explicit way to update col&row range. But now depends on the side effect of _movingHandler.
-            // call _movingHandler to update range, col selection, endRow should be last row of current sheet.
-            this._movingHandler(viewportPosX, viewportPosY, activeSelectionControl, rangeType);
-        }
-
-        this._setupPointerMoveListener(viewportMain, activeSelectionControl!, rangeType, scrollTimerType, viewportPosX, viewportPosY);
-
-        this._escapeShortcutDisposable = this._shortcutService.forceEscape();
-        this._scenePointerUpSub = scene.onPointerUp$.subscribeEvent(() => {
-            this._clearUpdatingListeners();
-            this._selectionMoveEnd$.next(this.getSelectionDataWithStyle());
-            this._escapeShortcutDisposable?.dispose();
-            this._escapeShortcutDisposable = null;
-        });
-    }
 
     /**
      * Init pointer move listener in each pointer down, unbind in each pointer up.
@@ -989,11 +859,11 @@ export class BaseSelectionRenderService extends Disposable implements ISheetSele
             (!evt.ctrlKey && !evt.shiftKey && !this._remainLastEnabled) ||
             (this._singleSelectionEnabled && !evt.shiftKey)
         ) {
-            this._clearSelectionControls();
+            this._clearAllSelectionControls();
         }
     }
 
-    private _performSelectionByTwoCells(
+    protected _performSelectionByTwoCells(
         currentCell: ISelectionCellWithCoord,
         startSelectionRange: IRangeWithCoord,
         skeleton: SpreadsheetSkeleton,
